groups created navbar content builder created for profile
