+++ Meetra/Views/Profile/Profile.swift
@@ -15,103 +15,86 @@ struct Profile: View {
     
     var body: some View {
         NavigationView {
-            ZStack {
+            
+            ProfileNavBar(title: NSLocalizedString("profile", comment: ""),
+                          showAlert: $profileVM.showAlert,
+                          message: profileVM.alertMessage, content: {
                 
-                if profileVM.loading {
-                    Loading()
-                } else if profileVM.profile != nil {
-                    ScrollView( showsIndicators: false ) {
-                        
-                        VStack( spacing: 25 ) {
-                            ProfileTopViewCompleteness(profile: profileVM.profile!)
+                ZStack {
+                    
+                    if profileVM.loading {
+                        Loading()
+                    } else if profileVM.profile != nil {
+                        ScrollView( showsIndicators: false ) {
                             
-                            
-                            if showFillPopup {
-                                ZStack( alignment: .topTrailing) {
-                                    Text( NSLocalizedString("fill_to_verify", comment: ""))
-                                        .kerning(0.24)
+                            VStack( spacing: 25 ) {
+                                ProfileTopViewCompleteness(profile: profileVM.profile!)
+                                
+                                
+                                if showFillPopup {
+                                    ZStack( alignment: .topTrailing) {
+                                        Text( NSLocalizedString("fill_to_verify", comment: ""))
+                                            .kerning(0.24)
+                                            .foregroundColor(.black)
+                                            .font(.custom("Inter-Regular", size: 12))
+                                            .lineSpacing(2)
+                                            .multilineTextAlignment(.center)
+                                            .padding(.horizontal, 25)
+                                            .padding(.vertical, 8)
+                                        
+                                        Button(action: {
+                                            withAnimation {
+                                                showFillPopup.toggle()
+                                            }
+                                        }, label: {
+                                            Image("close_button")
+                                        })
+                                    }.padding(8)
+                                        .background(AppColors.addProfileImageBG)
+                                        .cornerRadius(10)
+                                }
+                                
+                                VStack( alignment: .leading, spacing: 14) {
+                                    Text( NSLocalizedString("about", comment: ""))
+                                        .foregroundColor(.black)
+                                        .font(.custom("Inter-SemiBold", size: 18))
+                                    
+                                    Text( profileVM.profile!.bio )
                                         .foregroundColor(.black)
                                         .font(.custom("Inter-Regular", size: 12))
-                                        .lineSpacing(2)
-                                        .multilineTextAlignment(.center)
-                                        .padding(.horizontal, 25)
-                                        .padding(.vertical, 8)
                                     
-                                    Button(action: {
-                                        withAnimation {
-                                            showFillPopup.toggle()
-                                        }
-                                    }, label: {
-                                        Image("close_button")
-                                    })
-                                }.padding(8)
-                                    .background(AppColors.addProfileImageBG)
-                                    .cornerRadius(10)
-                            }
+                                }.padding(.horizontal, 30)
+                                    .frame(minWidth: 0,
+                                           maxWidth: .infinity,
+                                           alignment: .leading)
+                                
+                                VStack( alignment: .leading) {
+                                    Text( NSLocalizedString("interests", comment: ""))
+                                        .foregroundColor(.black)
+                                        .font(.custom("Inter-SemiBold", size: 18))
+                                    
+                                    TagsViewHelper(font: UIFont(name: "Inter-Regular", size: 12)!,
+                                                   parentWidth: UIScreen.main.bounds.size.width * 0.8,
+                                                   interests: profileVM.profile!.interests.map{UserInterestModel(same: true, name: String($0))})
+                                    
+                                }.padding(.horizontal, 30)
+                                    .frame(minWidth: 0,
+                                           maxWidth: .infinity,
+                                           alignment: .leading)
+                                    .padding(.bottom, UIScreen.main.bounds.size.height * 0.1)
+                                
+                            }.padding(.top, 25)
+                                .frame(minWidth: 0,
+                                       maxWidth: .infinity,
+                                       minHeight: 0,
+                                       maxHeight: .infinity)
                             
-                            VStack( alignment: .leading, spacing: 14) {
-                                Text( NSLocalizedString("about", comment: ""))
-                                    .foregroundColor(.black)
-                                    .font(.custom("Inter-SemiBold", size: 18))
-                                
-                                Text( profileVM.profile!.bio )
-                                    .foregroundColor(.black)
-                                    .font(.custom("Inter-Regular", size: 12))
-                                
-                            }.padding(.horizontal, 30)
-                            .frame(minWidth: 0,
-                                    maxWidth: .infinity,
-                                    alignment: .leading)
-                            
-                            VStack( alignment: .leading) {
-                                Text( NSLocalizedString("interests", comment: ""))
-                                    .foregroundColor(.black)
-                                    .font(.custom("Inter-SemiBold", size: 18))
-                                
-                                TagsViewHelper(font: UIFont(name: "Inter-Regular", size: 12)!,
-                                               parentWidth: UIScreen.main.bounds.size.width * 0.8,
-                                               interests: profileVM.profile!.interests                             .map{UserInterestModel(same: true, name: String($0))})
-                                
-                            }.padding(.horizontal, 30)
-                            .frame(minWidth: 0,
-                                    maxWidth: .infinity,
-                                    alignment: .leading)
-                            .padding(.bottom, UIScreen.main.bounds.size.height * 0.1)
-                            
-                        }.padding(.top, 25)
-                            .frame(minWidth: 0,
-                                   maxWidth: .infinity,
-                                   minHeight: 0,
-                                   maxHeight: .infinity)
-                        
-                    }.padding(.top, 1)
-                }
-                
-                
-            }.alert(isPresented: $profileVM.showAlert, content: {
-                Alert(title: Text("Error"), message: Text(profileVM.alertMessage), dismissButton: .default(Text("Got it!")))
-            })
-            .navigationBarTitle("", displayMode: .inline)
-            .navigationBarItems(leading: Text(NSLocalizedString("profile", comment: ""))
-                .kerning(0.56)
-                .foregroundColor(.black)
-                .font(.custom("Inter-Black", size: 28))
-                .padding(10), trailing: HStack( spacing: 20) {
-                    Button {
-                        
-                    } label: {
-                        Image("settings_icon")
+                        }.padding(.top, 1)
                     }
-                    
-                    Button {
-                        
-                    } label: {
-                        Image("icon_ring")
-                            .foregroundColor(.black)
-                    }
-                }).onAppear {
-                    profileVM.getProfile()
                 }
+            }).onAppear {
+                profileVM.getProfile()
+            }
         }.navigationViewStyle(StackNavigationViewStyle())
             .gesture(DragGesture().onChanged({ _ in
                 UIApplication.shared.endEditing()
